blockchain-project/requirements.txt
web3
py-solc-x
colorama
